# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend REST API for the Dermify application",
        "title": "Dermify API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/sessions/{id}/modules/ipl": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates an IPL (Intense Pulsed Light) module detail for a session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules-ipl"
                ],
                "summary": "Create IPL module",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "IPL module details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createIPLModuleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.IPLModuleDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sessions/{id}/modules/ipl/{moduleId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns an IPL module detail by module ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules-ipl"
                ],
                "summary": "Get IPL module",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Module ID",
                        "name": "moduleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.IPLModuleDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an IPL module detail. Requires version for optimistic locking.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules-ipl"
                ],
                "summary": "Update IPL module",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Module ID",
                        "name": "moduleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated IPL details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateIPLModuleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.IPLModuleDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apierrors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "AUTH_INVALID_CREDENTIALS"
                },
                "error": {
                    "type": "string",
                    "example": "invalid credentials"
                }
            }
        },
...
        "handlers.IPLModuleDetailResponse": {
            "type": "object",
            "properties": {
                "cooling_mode": {
                    "type": "string",
                    "example": "contact"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer",
                    "example": 1
                },
                "device_id": {
                    "type": "integer",
                    "example": 1
                },
                "filter_band": {
                    "type": "string",
                    "example": "560nm"
                },
                "fluence": {
                    "type": "number",
                    "example": 14.5
                },
                "handpiece_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "lightguide_size": {
                    "type": "string",
                    "example": "15x35mm"
                },
                "module_id": {
                    "type": "integer",
                    "example": 1
                },
                "notes": {
                    "type": "string"
                },
                "passes": {
                    "type": "integer",
                    "example": 2
                },
                "pulse_count": {
                    "type": "integer",
                    "example": 3
                },
                "pulse_delay": {
                    "type": "number",
                    "example": 30
                },
                "pulse_duration": {
                    "type": "number",
                    "example": 20
                },
                "total_pulses": {
                    "type": "integer",
                    "example": 120
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer",
                    "example": 1
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "handlers.createIPLModuleRequest": {
            "type": "object",
            "properties": {
                "cooling_mode": {
                    "type": "string"
                },
                "device_id": {
                    "type": "integer"
                },
                "filter_band": {
                    "type": "string"
                },
                "fluence": {
                    "type": "number"
                },
                "handpiece_id": {
                    "type": "integer"
                },
                "lightguide_size": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "passes": {
                    "type": "integer"
                },
                "pulse_count": {
                    "type": "integer"
                },
                "pulse_delay": {
                    "type": "number"
                },
                "pulse_duration": {
                    "type": "number"
                },
                "total_pulses": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.updateIPLModuleRequest": {
            "type": "object",
            "properties": {
                "cooling_mode": {
                    "type": "string"
                },
                "device_id": {
                    "type": "integer"
                },
                "filter_band": {
                    "type": "string"
                },
                "fluence": {
                    "type": "number"
                },
                "handpiece_id": {
                    "type": "integer"
                },
                "lightguide_size": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "passes": {
                    "type": "integer"
                },
                "pulse_count": {
                    "type": "integer"
                },
                "pulse_delay": {
                    "type": "number"
                },
                "pulse_duration": {
                    "type": "number"
                },
                "total_pulses": {
                    "type": "integer"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}